real-time-analytics/user-simulator/requirements.txt
kafka-python==2.0.2
faker==18.11.2
python-json-logger==2.0.7
requests==2.31.0
psycopg2-binary==2.9.7